@@ -15,59 +15,59 @@
     "lint:fix": "npm run lint:js -- --fix"
   },
   "dependencies": {
-    "@typegoose/typegoose": "^9.2.0",
+    "@apollo/server": "^4.9.5",
+    "@typegoose/typegoose": "^12.0.0",
     "@useoptic/express-middleware": "^0.0.5",
-    "agenda": "^3.1.0",
-    "agendash": "^2.1.1",
-    "apollo-server-core": "^3.5.0",
-    "apollo-server-express": "^3.5.0",
-    "argon2": "^0.27.0",
-    "body-parser": "^1.19.0",
-    "celebrate": "^13.0.3",
-    "class-validator": "^0.13.2",
+    "agenda": "^5.0.0",
+    "agendash": "^4.0.0",
+    "argon2": "^0.31.2",
+    "body-parser": "^1.20.2",
+    "celebrate": "^15.0.3",
+    "class-validator": "^0.14.0",
     "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
+    "dotenv": "^16.3.1",
     "errorhandler": "^1.5.1",
     "event-dispatch": "^0.4.1",
-    "eventemitter3": "^4.0.7",
-    "express": "^4.17.1",
-    "express-basic-auth": "^1.2.0",
-    "express-jwt": "^6.0.0",
-    "form-data": "^2.3.3",
-    "jsonwebtoken": "^8.5.1",
+    "eventemitter3": "^5.0.1",
+    "express": "^4.18.2",
+    "express-basic-auth": "^1.2.1",
+    "express-jwt": "^8.4.1",
+    "form-data": "^4.0.0",
+    "graphql": "^15.3.0",
+    "graphql-scalars": "^1.22.4",
+    "jsonwebtoken": "^9.0.2",
     "lodash": "^4.17.21",
-    "mailgun.js": "3.3.0",
+    "mailgun.js": "9.4.0",
     "method-override": "^3.0.0",
-    "moment": "^2.29.0",
-    "moment-timezone": "^0.5.31",
-    "mongoose": "^5.10.6",
+    "moment": "^2.30.1",
+    "moment-timezone": "^0.5.43",
+    "mongoose": "^8.0.3",
     "morgan": "^1.10.0",
-    "reflect-metadata": "^0.1.13",
+    "reflect-metadata": "^0.2.1",
     "type-graphql": "^1.1.1",
-    "typedi": "^0.8.0",
-    "winston": "^3.3.3"
+    "typedi": "^0.10.0",
+    "winston": "^3.11.0"
   },
   "devDependencies": {
-    "@types/agenda": "^2.0.9",
-    "@types/express": "^4.17.13",
-    "@types/jest": "^26.0.14",
-    "@types/lodash": "^4.14.161",
-    "@types/mongoose": "^5.7.36",
-    "@types/node": "^14.11.2",
-    "@typescript-eslint/eslint-plugin": "^4.2.0",
-    "@typescript-eslint/parser": "^4.2.0",
-    "eslint": "^7.9.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "jest": "^26.4.2",
-    "nodemon": "^2.0.4",
-    "prettier": "^2.1.2",
-    "ts-jest": "^26.4.0",
-    "ts-node": "^9.0.0",
-    "tsconfig-paths": "^3.11.0",
+    "@types/agenda": "^3.0.2",
+    "@types/express": "^4.17.21",
+    "@types/jest": "^29.5.11",
+    "@types/lodash": "^4.14.202",
+    "@types/node": "^20.10.5",
+    "@typescript-eslint/eslint-plugin": "^6.16.0",
+    "@typescript-eslint/parser": "^6.16.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.2",
+    "jest": "^29.7.0",
+    "nodemon": "^3.0.2",
+    "prettier": "^3.1.1",
+    "ts-jest": "^29.1.1",
+    "ts-node": "^10.9.2",
+    "tsconfig-paths": "^4.2.0",
     "tslint": "^5.20.1",
-    "typescript": "^4.0.3",
-    "winston-sentry-log": "^1.0.24"
+    "typescript": "^5.3.3",
+    "winston-sentry-log": "^1.0.26"
   },
   "bugs": {
     "url": "https://github.com/santiq/bulletproof-nodejs/issues"
